Rename classes -> plasma in codebase
