+++ plasmapy/plasma/tests/test_plasma_factory.py
@@ -3,13 +3,13 @@ import numpy as np
 import os
 import pytest
 
-import plasmapy.classes
+import plasmapy.plasma
 from plasmapy.particles.data.test import data_dir
 
 
 @pytest.fixture(scope="module")
 def h5(request):
-    h5 = plasmapy.classes.Plasma(hdf5=os.path.join(data_dir, "data00000255.h5"))
+    h5 = plasmapy.plasma.Plasma(hdf5=os.path.join(data_dir, "data00000255.h5"))
     yield h5
     h5.close()
 
@@ -17,18 +17,18 @@ def h5(request):
 class TestPlasma(object):
     def test_patters(self):
         # Input data whose specific subclass cannot be known
-        generic = plasmapy.classes.Plasma(blablobleh="spam")
-        assert isinstance(generic, plasmapy.classes.GenericPlasma)
+        generic = plasmapy.plasma.Plasma(blablobleh="spam")
+        assert isinstance(generic, plasmapy.plasma.GenericPlasma)
 
     def test_Plasma3D(self):
         # Input data for Plasma3D
-        three_dimensional = plasmapy.classes.Plasma(
+        three_dimensional = plasmapy.plasma.Plasma(
             domain_x=np.linspace(0, 1, 3) * u.m,
             domain_y=np.linspace(0, 1, 3) * u.m,
             domain_z=np.linspace(0, 1, 3) * u.m,
         )
-        assert isinstance(three_dimensional, plasmapy.classes.sources.Plasma3D)
-        assert isinstance(three_dimensional, plasmapy.classes.BasePlasma)
+        assert isinstance(three_dimensional, plasmapy.plasma.sources.Plasma3D)
+        assert isinstance(three_dimensional, plasmapy.plasma.BasePlasma)
 
     def test_PlasmaBlob(self):
         # Input data for PlasmaBlob
@@ -36,6 +36,6 @@ class TestPlasma(object):
         n_e = 1e26 * u.cm ** -3
         Z = 2.0
         particle = "p"
-        blob = plasmapy.classes.Plasma(T_e=T_e, n_e=n_e, Z=Z, particle=particle)
-        assert isinstance(blob, plasmapy.classes.sources.PlasmaBlob)
-        assert isinstance(blob, plasmapy.classes.BasePlasma)
+        blob = plasmapy.plasma.Plasma(T_e=T_e, n_e=n_e, Z=Z, particle=particle)
+        assert isinstance(blob, plasmapy.plasma.sources.PlasmaBlob)
+        assert isinstance(blob, plasmapy.plasma.BasePlasma)
